src/app/requirements.txt
pycryptodomex==3.20.2
pynacl===1.5.0
flask===2.3.2